refactorization of topopt examples
